libulb/catalog/requirements.txt
PyYaml
requests
BeautifulSoup4
html5lib
